Added functionality to mark movies as watched and renamed relevant classes for clarity
