+++ src/main/java/com/zggis/plextvtime/service/TVTimeServiceImpl.java
@@ -4,10 +4,10 @@ import com.zggis.plextvtime.exception.TVTimeException;
 import java.time.Duration;
 import java.util.HashMap;
 import java.util.Map;
-
 import com.zggis.plextvtime.util.ConsoleColor;
 import lombok.extern.slf4j.Slf4j;
 import org.javatuples.Triplet;
+import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.openqa.selenium.JavascriptExecutor;
@@ -102,24 +102,41 @@ public class TVTimeServiceImpl implements TVTimeService {
   }
 
   @Override
-  public String watchEpisode(String user, String episodeId) throws TVTimeException {
+  public String watchMedia(String user, String mediaId, String mediaType) throws TVTimeException {
     if (!isLoggedIn(user))
       throw new TVTimeException("You are not logged in");
     JSONObject responsePayload;
     WebClient client = getWebClient("https://app.tvtime.com");
     WebClient.UriSpec<WebClient.RequestBodySpec> uriSpec = client.post();
-    WebClient.RequestBodySpec bodySpec = uriSpec.uri(
-        "/sidecar?o=https://api2.tozelabs.com/v2/watched_episodes/episode/"
-            + episodeId
-            + "&is_rewatch=0");
-    WebClient.RequestHeadersSpec<?> requestHeadersSpec = bodySpec.bodyValue(userAuth.get(user).getValue2().toString());
-    requestHeadersSpec
-        .header(HttpHeaders.AUTHORIZATION, "Bearer " + userAuth.get(user).getValue0())
-        .header(
-            HttpHeaders.CONTENT_LENGTH,
-            String.valueOf(userAuth.get(user).getValue2().toString().getBytes().length))
-        .header(HttpHeaders.HOST, "app.tvtime.com:80")
-        .retrieve();
+    WebClient.RequestBodySpec bodySpec;
+    WebClient.RequestHeadersSpec<?> requestHeadersSpec;
+    String movieUUID = "null";
+    if (mediaType.equals("movie")) {
+      try {
+        movieUUID = getMovieUUID(user, mediaId);
+      } catch (JSONException e) {
+        log.info("Movie not found in TVTime database");
+      }
+      if (movieUUID.equals("")) {
+        log.warn("Movie not found in TVTime database");
+      }
+      log.debug("Movie UUID: {}", movieUUID);
+      bodySpec = uriSpec.uri("/sidecar?o=https://msapi.tvtime.com/prod/v1/tracking/" + movieUUID + "/watch");
+      // For movies, we must use a different endpoint and remove the jwt_refresh_token
+      // from the body
+      requestHeadersSpec = bodySpec.header(HttpHeaders.AUTHORIZATION, "Bearer " + userAuth.get(user).getValue0())
+          .header(HttpHeaders.HOST, "app.tvtime.com:80");
+    } else {
+      bodySpec = uriSpec
+          .uri("/sidecar?o=https://api2.tozelabs.com/v2/watched_episodes/episode/" + mediaId + "&is_rewatch=0");
+      requestHeadersSpec = bodySpec.bodyValue(userAuth.get(user).getValue2().toString());
+      requestHeadersSpec
+          .header(HttpHeaders.AUTHORIZATION, "Bearer " + userAuth.get(user).getValue0())
+          .header(HttpHeaders.CONTENT_LENGTH,
+              String.valueOf(userAuth.get(user).getValue2().toString().getBytes().length))
+          .header(HttpHeaders.HOST, "app.tvtime.com:80");
+    }
+
     Mono<String> response = requestHeadersSpec.retrieve().bodyToMono(String.class);
     try {
       responsePayload = new JSONObject(response.block());
@@ -127,7 +144,7 @@ public class TVTimeServiceImpl implements TVTimeService {
     } catch (JSONException e) {
       log.error(e.getMessage(), e);
     }
-    return null;
+    return "";
   }
 
   private boolean isLoggedIn(String user) {
@@ -140,5 +157,33 @@ public class TVTimeServiceImpl implements TVTimeService {
         .defaultHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
         .codecs(configurer -> configurer.defaultCodecs().maxInMemorySize(16 * 1024 * 1024))
         .build();
+  }
+
+  public String getMovieUUID(String user, String movieId) {
+    JSONObject responsePayload;
+    WebClient client = getWebClient("https://app.tvtime.com");
+    Mono<String> response = client.get()
+        .uri("/sidecar?o=https://search.tvtime.com/v1/search/series,movie&q=" + movieId + "&offset=0&limit=1")
+        .header(HttpHeaders.AUTHORIZATION, "Bearer " + userAuth.get(user).getValue0())
+        .header(HttpHeaders.HOST, "app.tvtime.com:80")
+        .retrieve()
+        .bodyToMono(String.class);
+
+    try {
+      responsePayload = new JSONObject(response.block());
+      if (responsePayload.getString("status").equals("success")) {
+        JSONArray dataArray = responsePayload.getJSONArray("data");
+        for (int i = 0; i < dataArray.length(); i++) {
+          JSONObject movie = dataArray.getJSONObject(i);
+          int movieIdInt = Integer.parseInt(movieId);
+          if (movie.getInt("id") == movieIdInt) {
+            return movie.getString("uuid");
+          }
+        }
+      }
+    } catch (JSONException e) {
+      log.error(e.getMessage(), e);
+    }
+    return null;
   }
 }
